修bug

+++ debian/update_ui.sh
@@ -21,7 +21,7 @@ ensure_directories() {
 
 # 清理临时文件
 cleanup() {
-    rm -rf "$TEMP_DIR"/*
+    rm -rf "${TEMP_DIR:?}"/*
 }
 
 # 从配置文件获取下载URL
@@ -84,7 +84,7 @@ update_ui() {
     if unzip_with_busybox "$TEMP_DIR/ui.zip" "$TEMP_DIR"; then
         # 移动文件到UI目录
         sudo mv "$TEMP_DIR"/*/* "$UI_DIR"
-        echo -e "${CYAN}控制面板UI更新完成！${NC}"
+        echo -e "${CYAN}控制面板UI更新完成,请刷新浏览器！${NC}"
     else
         # 还原备份
         restore_ui
